requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
celery==5.3.1
redis==5.0.1
python-decouple==3.8
Pillow==10.0.1
dj-database-url==2.1.0
psycopg2-binary==2.9.7